refactor: move convert_bounding_box_absolute_to_relative to utils.image
and add unit tests
